fix code typos mentioned in https://github.com/microsoft/SoftTeacher/pull/5 and refactor
