+++ ssod/datasets/pipelines/__init__.py
@@ -1,2 +1,2 @@
-from .formating import *
+from .formatting import *
 from .rand_aug import *
